refactor: api 세분화
