+++ src/api/cart.ts
@@ -1,10 +1,5 @@
 import { api } from 'api';
 
-export const getProductList = async (server: string) => {
-  const data = await api.get(`${server}/products`);
-  return data;
-};
-
 export const getCartList = async (server: string) => {
   const data = await api.get(`${server}/cart-items`);
   return data;
