+++ src/api/products.ts
@@ -1,0 +1,6 @@
+import { api } from 'api';
+
+export const getProductList = async (server: string) => {
+  const data = await api.get(`${server}/products`);
+  return data;
+};
